Renaming contenthash to deephash
